% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rows.R
\name{filter_rows}
\alias{filter_rows}
\title{filter_rows}
\usage{
filter_rows(data, keep_only_rows_with_x_valid_ratio = 0.5)
}
\arguments{
\item{data}{lowest-level data frame}

\item{keep_only_rows_with_x_valid_ratio}{only keep rows with more than this ratio of non-NA-values}
}
\value{
Filtered lowest-level data frame.
}
\description{
Filtering out rows that do not contain more real values than the specified ratio.
The ratio symbolizes an exclusive lower bound.
Exceptions: ratio = 0 means no filtering is done, ratio = 1 means keeping only rows with 100 percent real values.
}
