% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2_transform.R
\name{log2_transform}
\alias{log2_transform}
\title{log2_transform}
\usage{
log2_transform(intensities)
}
\arguments{
\item{intensities}{lowest-level data frame}
}
\value{
log2 transformed lowest-level data frame
}
\description{
log2_transform
}
