% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_normalize.R
\name{median_normalize}
\alias{median_normalize}
\title{median_normalize}
\usage{
median_normalize(data)
}
\arguments{
\item{data}{lowest-level data frame}
}
\value{
median normalized lowest-level data frame
}
\description{
Median normalization of the lowest-level data frame
}
