% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_row.R
\name{normalize_row}
\alias{normalize_row}
\title{normalize_row}
\usage{
normalize_row(
  intensities,
  exp_design,
  active = FALSE,
  ref = NULL,
  performTotalSum = F,
  refFunc = median,
  na.rm = TRUE
)
}
\arguments{
\item{intensities}{lowest-level data frame}

\item{exp_design}{experimental design data frame}

\item{active}{set as TRUE for interactive mode}

\item{ref}{references can be directly set here}

\item{performTotalSum}{set as TRUE for total sum scaling}

\item{refFunc}{can be set as median or sum}

\item{na.rm}{tells whether NA values are removed}
}
\value{
row-wise normalized lowest-level data frame
}
\description{
Row-wise normalization
}
