% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_row_ref.R
\name{normalize_row_ref}
\alias{normalize_row_ref}
\title{normalize_row_ref}
\usage{
normalize_row_ref(
  lowest_level_df,
  exp_design,
  ref,
  refFunc = median,
  na.rm = TRUE
)
}
\arguments{
\item{lowest_level_df}{lowest-level data frame}

\item{exp_design}{experimental design data frame}

\item{ref}{vector of references}

\item{refFunc}{can be set as median or sum}

\item{na.rm}{tells whether NA values are removed}
}
\value{
row-wise normalized lowest-level data frame
}
\description{
Helper function to perform row-wise normalization
}
