% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.smooth2.R
\name{panel.smooth2}
\alias{panel.smooth2}
\title{panel.smooth2}
\usage{
panel.smooth2(
  x,
  y,
  col = par("col"),
  bg = NA,
  pch = ".",
  cex = 1,
  col.smooth = "red",
  span = 2/3,
  iter = 3,
  ...
)
}
\arguments{
\item{x}{numeric vector of the same length as y}

\item{y}{numeric vector of the same length as x}

\item{col}{numeric or character code for the color(s) of the points}

\item{bg}{background color for symbol used for the points}

\item{pch}{symbol used for the points}

\item{cex}{expansion factor used for the points}

\item{col.smooth}{color to be used by lines function for drawing the smooths}

\item{span}{smoothing parameter f for lowess function}

\item{iter}{number of robustness iterations for lowess function}

\item{...}{further arguments}
}
\value{
Lower panel (scatter plots) for pairs.
}
\description{
Panel function used in the pairs plot (correlation plot) as lower.panel.
}
