% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot2.R
\name{pcaPlot2}
\alias{pcaPlot2}
\title{pcaPlot2}
\usage{
pcaPlot2(
  data,
  exp_design,
  show_labels = F,
  pdf_mode = F,
  set_colors = NULL,
  set_symbols = NULL
)
}
\arguments{
\item{data}{lowest-level data frame}

\item{exp_design}{experimental design data frame}

\item{show_labels}{Boolean that states whether the labels for the data points are shown}

\item{pdf_mode}{Boolean that can be set as T to adjust the plot margins for the PDF}

\item{set_colors}{optionally set colors for the batches as a vector of Strings}

\item{set_symbols}{optionally set symbols for the conditions as a vector of numerics}
}
\value{
The returned value is the score of the PCA plot. The PCA plot is displayed.
}
\description{
PCA plot including the first and second principal components.The data points are colored by batch, while their symbols represent the conditions.
In addition, the score of the PCA plot is returned. The score describes how strictly the data points are clustered by condition.
}
