% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlations.R
\name{plot_correlations}
\alias{plot_correlations}
\title{plot_correlations}
\usage{
plot_correlations(data)
}
\arguments{
\item{data}{lowest-level data frame}
}
\value{
Displays a correlation plot.
}
\description{
Correlation plot generated by the function pairs.
The plot includes a Pearson correlation coefficient matrix and scatter plots
showing pairwise relationships of the variables.
}
