% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{plot_results}
\usage{
plot_results(lowest_level_df, main = "", output_dir = "")
}
\arguments{
\item{lowest_level_df}{lowest-level data frame}

\item{main}{optionally specify the name of the output file and title for the plots}

\item{output_dir}{optionally specify the directory in which the output file will be saved}
}
\value{
Creates an output pdf file containing the generated plots
}
\description{
Function to be called to generate all plots and save them inside an output pdf file
}
