% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStats.R
\name{plotStats}
\alias{plotStats}
\title{plotStats}
\usage{
plotStats(data, exp_design, main = "", show_labels = T, pdf_mode = F)
}
\arguments{
\item{data}{lowest-level data frame}

\item{exp_design}{experimental design data frame}

\item{main}{title for the plots}

\item{show_labels}{states whether the labels for the data points are shown inside the PCA plot}

\item{pdf_mode}{used for adjusting the PCA plot for the PDF when generated in plot_results}
}
\value{
creates correlation plot, heatmap, component plot, and PCA plot
}
\description{
Generates plots
}
