% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_files}
\alias{read_files}
\title{read_files}
\usage{
read_files(
  file1,
  file2,
  rm_only_by_site = TRUE,
  rm_reverse = TRUE,
  rm_contaminant = TRUE
)
}
\arguments{
\item{file1}{proteinGroups.txt file}

\item{file2}{file of the experimental design}

\item{rm_only_by_site}{remove rows being positive for only by site}

\item{rm_reverse}{remove rows being positive for reverse}

\item{rm_contaminant}{remove rows being positive for a contaminant}
}
\value{
A list storing the lowest-level data frame, the data frame for the experimental design
and a data frame with the remaining columns if present
}
\description{
Reading of the input files
}
