% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_common_prefix.R
\name{remove_common_prefix}
\alias{remove_common_prefix}
\title{remove_common_prefix}
\usage{
remove_common_prefix(strings)
}
\arguments{
\item{strings}{set of strings}
}
\value{
list of the common prefix and the set of strings without the common prefix
}
\description{
Helper function to remove the common prefix of a set of strings
}
