% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{run_app}
\usage{
run_app(local = FALSE)
}
\arguments{
\item{local}{Boolean that states whether the app is called locally. If set, further options for manual downloads are included in the app.}
}
\value{
executes the shiny application
}
\description{
Execution of the shiny application. The application includes the performance of row-wise normalization,
as well as the total sum normalization, VST, VSN, quantile normalization, ComBat, and M-ComBat.
}
