% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.normalize.R
\name{sum.normalize}
\alias{sum.normalize}
\title{sum.normalize}
\usage{
\method{sum}{normalize}(data, refFunc = sum, norm = T, na.rm = T)
}
\arguments{
\item{data}{lowest-level data frame}

\item{refFunc}{default is sum}

\item{norm}{scaling of values}

\item{na.rm}{tells whether NA values are removed}
}
\value{
scaled lowest-level data frame
}
\description{
Total sum normalization of the lowest-level data frame
}
