% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_normalize.R
\name{sum_normalize}
\alias{sum_normalize}
\title{sum_normalize}
\usage{
sum_normalize(data, refFunc = "sum", norm = T, na.rm = T)
}
\arguments{
\item{data}{lowest-level data frame}

\item{refFunc}{String that is to be set as "sum"}

\item{norm}{scaling of values}

\item{na.rm}{Boolean that tells whether NA values are removed}
}
\value{
scaled lowest-level data frame
}
\description{
Total sum normalization of the lowest-level data frame
}
