% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_outfile2.R
\name{write_outfile2}
\alias{write_outfile2}
\title{write_outfile}
\usage{
write_outfile2(
  lowest_level_df,
  additional_cols = NULL,
  filename = "",
  output_dir = ""
)
}
\arguments{
\item{lowest_level_df}{lowest-level data frame}

\item{additional_cols}{optionally append the additional columns from the input file}

\item{filename}{optionally specify the name of the output file}

\item{output_dir}{optionally specify the directory in which the output file will be saved}
}
\value{
Creating an output file
}
\description{
Write the normalized values into a file
}
