% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2_transform.R
\name{log2_transform}
\alias{log2_transform}
\title{log2_transform}
\usage{
log2_transform(lowest_level_df)
}
\arguments{
\item{lowest_level_df}{lowest-level data frame}
}
\value{
Log2 transformed lowest-level data frame.
}
\description{
Logarithmic transformation. This represents as an optional pre-processing step.
}
