% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_normalize.R
\name{median_normalize}
\alias{median_normalize}
\title{median_normalize}
\usage{
median_normalize(data)
}
\arguments{
\item{data}{lowest-level data frame}
}
\value{
Median normalized lowest-level data frame.
}
\description{
Median normalization. This represents as an optional pre-processing step.
NA values are ignored inside the calculation of the median.
}
