% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_row.R
\name{normalize_row}
\alias{normalize_row}
\title{normalize_row}
\usage{
normalize_row(
  lowest_level_df,
  exp_design,
  active = FALSE,
  ref = NULL,
  refFunc = "median",
  na.rm = TRUE
)
}
\arguments{
\item{lowest_level_df}{lowest-level data frame}

\item{exp_design}{experimental design data frame}

\item{active}{Boolean that can be set as TRUE for interactive mode}

\item{ref}{references can be set here as Strings inside a vector}

\item{refFunc}{String that can be set as "median" or "sum"}

\item{na.rm}{Boolean that tells whether NA values are removed}
}
\value{
row-wise normalized lowest-level data frame
}
\description{
Row-wise normalization of the lowest-level data
}
