% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_row_ref.R
\name{normalize_row_ref}
\alias{normalize_row_ref}
\title{normalize_row_ref}
\usage{
normalize_row_ref(
  lowest_level_df,
  exp_design,
  ref,
  refFunc = "median",
  na.rm = TRUE
)
}
\arguments{
\item{lowest_level_df}{lowest-level data frame}

\item{exp_design}{experimental design data frame}

\item{ref}{the reference conditions as Strings inside a vector}

\item{refFunc}{String that can be set as "median" or "sum" to define the reference function inside the calculation}

\item{na.rm}{Boolean that tells whether NA values are removed inside the reference function}
}
\value{
row-wise normalized lowest-level data frame
}
\description{
Helper function to perform row-wise normalization. References must be specified.
}
