% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.cor.R
\name{panel.cor}
\alias{panel.cor}
\title{panel.cor}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor, ...)
}
\arguments{
\item{x}{numeric vector, the first variable, provided by pairs}

\item{y}{numeric vector, the second variable, provided by pairs}

\item{digits}{number of digits to use on the pairs plot for the correlations}

\item{prefix}{text to prefix the correlation coefficients}

\item{cex.cor}{character expansion factor}

\item{...}{further arguments}
}
\value{
Upper panel (correlation coefficients) for pairs.
}
\description{
Panel function used in the pairs plot (correlation plot) as upper.panel.
}
