% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.smooth2.R
\name{panel.smooth2}
\alias{panel.smooth2}
\title{panel.smooth2}
\usage{
panel.smooth2(
  x,
  y,
  col = par("col"),
  bg = NA,
  pch = ".",
  cex = 1,
  col.smooth = "red",
  span = 2/3,
  iter = 3,
  ...
)
}
\description{
Helper function for generating plots
}
