% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.R
\name{pcaPlot}
\alias{pcaPlot}
\title{pcaPlot}
\usage{
pcaPlot(data)
}
\arguments{
\item{data}{lowest-level data frame}
}
\value{
creates a component plot
}
\description{
Helper function for generating the component plot
}
