% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.R
\name{pcaPlot}
\alias{pcaPlot}
\title{pcaPlot}
\usage{
pcaPlot(data)
}
\arguments{
\item{data}{lowest-level data frame}
}
\value{
displays a component plot
}
\description{
Component plot generated by the function biplot. The plot includes the first and second principal components.
Rows containing NA values are not included.
}
