% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot2.R
\name{pcaPlot2}
\alias{pcaPlot2}
\title{pcaPlot2}
\usage{
pcaPlot2(data, exp_design, main = "", show_labels = F, pdf_mode = F)
}
\arguments{
\item{data}{lowest-level data frame}

\item{exp_design}{experimental design data frame}

\item{main}{title for the plot}

\item{show_labels}{states whether the labels for the data points are shown}

\item{pdf_mode}{used for adjusting the plot for the PDF when generated in plot_results}
}
\value{
creates a PCA plot
}
\description{
Helper function for generating PCA plot with principal components
}
