% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlations.R
\name{plot_correlations}
\alias{plot_correlations}
\title{plot_correlations}
\usage{
plot_correlations(data)
}
\arguments{
\item{data}{lowest-level data frame}
}
\value{
creates a correlation plot
}
\description{
Helper function for generating the correlation plot
}
