% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(data, main = "")
}
\arguments{
\item{data}{lowest-level data frame}

\item{main}{title for the plot}
}
\value{
creates a heatmap plot
}
\description{
Helper function for generating the heatmap plot
}
