% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(data, exp_design, batch_colors = NULL)
}
\arguments{
\item{data}{lowest-level data frame}

\item{exp_design}{experimental design data frame}

\item{batch_colors}{optionally set colors for the batches as a vector of Strings}
}
\value{
Displays a heatmap plot.
}
\description{
Heatmap plot generated by the function pheatmap. The legends show the respective batch and condition of a variable.
}
