% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{plot_results}
\usage{
plot_results(
  lowest_level_df,
  exp_design,
  filename = "",
  output_dir = "",
  show_labels = F,
  svg = F,
  set_colors = NULL,
  set_symbols = NULL
)
}
\arguments{
\item{lowest_level_df}{lowest-level data frame}

\item{exp_design}{experimental design data frame}

\item{filename}{optionally specify the name of the output file}

\item{output_dir}{optionally specify the directory in which the output file will be saved}

\item{show_labels}{states whether the labels for the data points are shown inside the PCA plot}

\item{svg}{states whether a svg file is created for each of the plots additionally}

\item{set_colors}{optionally set colors for the batches inside the PCA plot and inside the heatmap}

\item{set_symbols}{optionally set symbols for the conditions inside the PCA plot}
}
\value{
Creates an output pdf file containing the generated plots
}
\description{
Function to be called to generate all plots and save them inside an output pdf file
}
