% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStats.R
\name{plotStats}
\alias{plotStats}
\title{plotStats}
\usage{
plotStats(data, main = "")
}
\arguments{
\item{data}{lowest-level data frame}

\item{main}{title for the plots}
}
\value{
creates correlation plot, heatmap, component plot, and PCA plot
}
\description{
Generates plots
}
