% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_files}
\alias{read_files}
\title{read_files}
\usage{
read_files(
  data,
  design,
  rm_only_by_site = TRUE,
  rm_reverse = TRUE,
  rm_contaminant = TRUE
)
}
\arguments{
\item{data}{file of the data}

\item{design}{file of the experimental design}

\item{rm_only_by_site}{Boolean that states whether rows being positive for only by site are removed}

\item{rm_reverse}{Boolean that states whether rows being positive for reverse are removed}

\item{rm_contaminant}{Boolean that states whether rows being positive for a contaminant are removed}
}
\value{
List storing the lowest-level data frame, the data frame for the experimental design,
 a data frame with the remaining columns if present, a vector of colors that can be used for the batches in the PCA plot and the heatmap plot,
 and a vector of symbols that can be used for the conditions in the PCA plot.
}
\description{
Reading of the input files.
}
