% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_outfile.R
\name{write_outfile}
\alias{write_outfile}
\title{write_outfile}
\usage{
write_outfile(lowest_level_df, additional_cols = NULL, filename = "")
}
\arguments{
\item{lowest_level_df}{lowest-level data frame}

\item{additional_cols}{append the additional columns from the input file}

\item{filename}{can be optionally specified}
}
\value{
Creating an output file
}
\description{
Write the normalized values into a file
}
