% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_outfile.R
\name{write_outfile}
\alias{write_outfile}
\title{write_outfile}
\usage{
write_outfile(
  lowest_level_df,
  additional_cols = NULL,
  filename = "",
  output_dir = ""
)
}
\arguments{
\item{lowest_level_df}{lowest-level data frame}

\item{additional_cols}{optionally append the additional columns from the input file as a data frame}

\item{filename}{optionally specify the name of the output file}

\item{output_dir}{optionally specify the directory in which the output file will be saved}
}
\value{
Creates and downloads an output file
}
\description{
Download of the normalized values written into a file with comma-separated values.
}
